requirements.txt
python-dotenv
google-generativeai
streamlit
Pillow
python-docx
pypdf2
openpyxl
pandas